backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pytest==8.3.2
httpx==0.27.2
python-multipart==0.0.9
# Upgrade SQLAlchemy for Python 3.13 compatibility
sqlalchemy==2.0.31
# psycopg2-binary should work; if build issues persist on Py3.13 consider installing build tools or switching to psycopg (psycopg==3.x)
psycopg2-binary==2.9.9
alembic==1.13.1

